% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA2_ICC_AD_Index.R
\name{AD_index}
\alias{AD_index}
\alias{AD_Index}
\alias{AD_Index2}
\alias{AD_Index.data.frame}
\alias{AD_Index.formula}
\alias{AD_Index.default}
\title{Hypothese Jeder Judge beurteilt das Item gleich}
\usage{
AD_Index(x, ...)

AD_Index2(x, data, ..., A = 5, type = "judge",
  caption = "Interrater Agreement",
  note = "AD index (based on deviations from the item medians)")

\method{AD_Index}{data.frame}(data, ...)

\method{AD_Index}{formula}(x, data, type = "judge", ...)

\method{AD_Index}{default}(x, ...)
}
\arguments{
\item{x}{Objekt Formula oder Vektor}

\item{...}{weitere Objekte
na.rm normalerweise bna.rm==TRUE soll nicht geaendert werden}

\item{data}{Objekt}

\item{A}{A=5 Anzahl stufen der  Likertskala}

\item{type}{Lang oder Breit type = "judge" oder "item}

\item{caption}{caption ="Interrater Agreement",}

\item{note}{Note}
}
\value{
Ein data.frame
}
\description{
Estimating Interrater Agreement With the
Average Deviation Index: A User’s Guide MICHAEL J. BURKE WILLIAM P. DUNLAP
Basiert auf 642 Raphaela Willitsch
wo Estimating Interrater Agreement With the Average Deviation Index: A User’s Guide MICHAEL J. BURKE
WILLIAM P. DUNLAP
die Methode beschreibt.

Estimating within-group interrater reliability with and without response bias.
By James, Lawrence R.; Demaree, Robert G.; Wolf, Gerrit
Journal of Applied Psychology, Vol 69(1), Feb 1984, 85-98.
Abstract
Presents methods for assessing agreement among the judgments made by a single group of judges on a single variable in regard to a single target. For example, the group of judges could be editorial consultants, members of an assessment center, or members of a team. The single target could be a manuscript, a lower level manager, or a team. The variable on which the target is judged could be overall publishability in the case of the manuscript, managerial potential for the lower level manager, or a team cooperativeness for the team. The methods presented are based on new procedures for estimating interrater reliability. For such situations, these procedures furnish more accurate and interpretable estimates of agreement than estimates provided by procedures commonly used to estimate agreement, consistency, or interrater reliability. The proposed methods include processes for controlling for the spurious influences of response biases (e.g., positive leniency and social desirability) on estimates of interrater reliability. (49 ref) (PsycINFO Database Record (c) 2016 APA, all rights reserved)

Hypothese Jeder Judge beurteilt das Item gleich.

rwg: oeherer Wertbessere Uebereinstimmung
}
\examples{
# data
#  Item     x1 x2 x3 x4 x5 x6
# Judge 1   3  4  3  2  4  3
# Judge 2   2  3  3  2  4  4
# Judge 3   4  3  4  3  4  3
# Judge 4   3  3  2  2  2  4
# Judge 5   3  2  2  4  2  3
# Judge 6   4  2  4  3  2  3
# Judge 7   2  3  2  3  3  4
# Judge 8   3  4  4  3  3  2
# Judge 9   4  2  3  4  3  2
# Judge 10  2  4  3  4  3  2
}
