% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R, R/corr.R, R/icc.R, R/likert.R, R/pca.R,
%   R/reliability.R, R/rsquer.R, R/t_test.R, R/xtabs.R
\name{APA_}
\alias{APA_}
\alias{APA_NULL}
\alias{APA_Correlation}
\alias{APA_ICC}
\alias{ICC2}
\alias{APA_Likert}
\alias{APA_PCA}
\alias{APA_Reliability}
\alias{APA_R2}
\alias{APA_Ttest}
\alias{APA.xtabs}
\alias{APA_Xtabs}
\title{APA Syle Table}
\usage{
APA_NULL(x, ...)

APA_Correlation(x, ..., caption = "Korrelation", note = NULL)

APA_ICC(x, ..., caption = "ICC", type = c(1, 4))

ICC2(x, ..., caption = "ICC", type = c(1, 4))

APA_Likert(...)

APA_PCA(data, ..., include.pca = TRUE, include.plot = FALSE,
  include.kmo = TRUE, w = 5, h = 5,
  caption = "Standardized loadings (pattern matrix) based upon correlation matrix",
  note = "", main = "")

APA_Reliability(..., caption = "", note = "")

APA_R2(..., caption, note)

APA_Ttest(x, data, var.equal = FALSE, paired = FALSE,
  alternative = "two.sided", include.mean = TRUE, include.d = TRUE,
  include.mean.diff = TRUE, include.se = TRUE, type = "t.test",
  digits = 2, ...)

\method{APA}{xtabs}(x, ...)

APA_Xtabs(x, data, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy data.frame}

\item{...}{extra arguments}

\item{caption, note}{Ueberschrift}

\item{note}{Note}

\item{type}{Methode t.test oder wilcox.test}

\item{data}{Data.frame}

\item{include.pca, include.plot, include.kmo}{type=c("pca", "plot", "kmo")}

\item{w, h, main}{Grafik Fenster w=5 h=5 Ueberschrift}

\item{var.equal, paired, alternative}{an t.Test var.equal = FALSE  "two.sided"}

\item{include.mean}{Mittelwerte mit SD}

\item{include.d}{Cohens d}

\item{include.mean.diff}{Mittlere Differenzen}

\item{include.se}{Standardfehler noch nicht Fertig}

\item{digits}{Nachkommastellen}

\item{caption}{ueberschrift}

\item{x}{Formula}

\item{...}{an Output}
}
\value{
a data.frame

list
}
\description{
All functions in stringr start with APA_ and take a Object (formula, data.frame, lm, ...) 
as the first argument.

APA_Correlation: Korrelationstabelle (Interkorrelationen mit hilfe
der Funktion \code{\link[Hmisc]{rcorr}}. Erlaubt ist die getrennte Auswertung ueber groups bzw mit by.
Interne Funktion \code{\link{corr_tabel}}.

APA_ICC  Intra-Klassen-Korrelation ( mit psych ).

APA_PCA: Faktoranalyse wie SPSS \link{PCA}

APA_R2: R2-Tabelle (noch nicht fretig)

APA_Ttest: Berechnung von Mittelwerten und Mittelwertdifferenzen und des T-Test

APA.xtabs Chi-Quadrat aus Kreuztabellen

APA_Xtabs Kreuztabellen
}
\examples{
n<- 2*8
e<- rnorm(n)*10
data<- data.frame(a=rnorm(n) + e,
                  b=rnorm(n), c=rnorm(n), d=rnorm(n) + e,
                  g=gl(2, 8, labels = c("Control", "Treat")))
#
APA_Correlation(~a+b+c, data, caption="~a+b+c")

APA_Correlation(~a+b+c, data, caption="mit Output: ~a+b+c")


 # library(arm)
 # windows(7,7)
 # corrplot(data[Cs(a,b,c,d)], abs=TRUE, n.col.legend=7)

 # install.packages("PerformanceAnalytics")

 #library("PerformanceAnalytics")
 #?chart.Correlation#(decathlon2.active[, 1:6], histogram=TRUE, pch=19)
 #data(managers)
 #chart.Correlation(managers[,1:8], histogram=TRUE, pch="+")

 require(stpvers)
 APA_Reliability(~F3+F2+F10+F11, fkv, check.keys =TRUE)

fit1<-lm(score ~ grade + treatment, schools)
fit2<-lm(score ~ grade + treatment + stdTest, schools)
APA_R2(fit1, fit2)


#require(stpvers)

#Projekt("html")

APA_Ttest(m1+m2 ~ geschl, varana)
varanax<-Melt2(m1+m2~nr,varana , key="time", value="m")

# broom::tidy(with(varana, t.test(m1,m2 ) ))
# broom::tidy(  t.test(m~time, varanax, var.equal=FALSE)) )

APA_Ttest(m~time, varanax, paired = TRUE)

APA_Ttest(m~time, varanax, var.equal=TRUE)
APA_Ttest(m~time, varanax, var.equal=FALSE)
#broom::tidy(t.test(m~time, varanax, var.equal=TRUE))
#broom::tidy(t.test(m~time, varanax, var.equal=FALSE))

#End()
}
