% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA2_Effsize.R
\name{APA_Effsize}
\alias{APA_Effsize}
\title{APA_Effsize}
\usage{
APA_Effsize(x, ...)
}
\arguments{
\item{x}{Objekt oder Formel}

\item{...}{weitere Optionen}

\item{type}{bis jetzt nur cohens d}
}
\description{
Cohen's d and Hedges g effect size
Between groups
Cohen's d Small 0.2, Medium 0.5, Large 0.8, Very large 1.3
Odds ratio (OR) Small 1.5, Medium 2, Large 3
Relative risk or risk ratio (RR) R Small 2, Medium 3, Large 4
}
\details{
Measures of association
Pearson's r correlation Small 0.2,  Medium 0.5, Large 0.8
r2 coefficient of determination Small 0.04, Medium 0.25, Large 0.64
Quelle: http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3444174/pdf/i1949-8357-4-3-279.pdf

Gestolen von https://cran.r-project.org/web/packages/effsize/effsize.pdf
}
