% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression1.R
\name{APA_Validation}
\alias{APA_Validation}
\title{Validation of Linear Models Assumptions}
\usage{
APA_Validation(..., include.F = TRUE, include.heteroskedasticity = TRUE,
  include.durbin = TRUE, include.levene = FALSE, include.bartlett = FALSE,
  caption = "Testing Linear Regression Models", note = "", names = NULL)
}
\arguments{
\item{...}{}

\item{include.F}{}

\item{include.heteroskedasticity}{}

\item{include.durbin}{}

\item{include.levene}{}

\item{include.bartlett}{}

\item{caption}{}

\item{note}{}

\item{names}{}
}
\value{
Output als html und data.frame
}
\description{
Test von Heteroskedasticity und Autocorrelation
}
