% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R, R/mean2.R
\name{Berechne}
\alias{Berechne}
\alias{berechne}
\alias{berechne.default}
\alias{berechne.data.frame}
\alias{berechneMean}
\alias{mean2}
\alias{median2}
\alias{sd2}
\alias{CI}
\alias{StdErr}
\alias{Mean2}
\alias{Mean2.formula}
\alias{Mean2.default}
\alias{Median2}
\alias{Median2.formula}
\alias{Median2.default}
\alias{Meanci2}
\alias{Prozent2APA}
\alias{Mean2default}
\alias{Median2default}
\alias{Prozent2default}
\title{Berechne Mittelwerte}
\usage{
berechne(...)

\method{berechne}{default}(data, x, by = "1", measure, type = 2,
  fun = NULL, fm = NULL, digits = NULL, measure.name = NULL)

\method{berechne}{data.frame}(.data, ..., by = "1", type = 1,
  na.rm = TRUE, conf.interval = 0.95, .drop = TRUE)

berechneMean(data = NULL, measurevar, by = NULL, na.rm = TRUE,
  conf.interval = 0.95, .drop = TRUE)

mean2(x, na.rm = TRUE, ...)

median2(x, na.rm = TRUE, ...)

sd2(x, na.rm = TRUE, ...)

CI(x, ci = 0.95, na.rm = TRUE, ...)

StdErr(x, ...)

Mean2(x, ...)

\method{Mean2}{formula}(x, data, ...)

\method{Mean2}{default}(x, digits = NULL)

Median2(x, ...)

\method{Median2}{formula}(x, data, ...)

\method{Median2}{default}(x, digits = NULL,
  median.style = get_my_options()$apa.style$mittelwert$median.style, ...)

Meanci2(x, digits = NULL, ...)

Prozent2APA(x, exclude = NA, digits = 1, max_factor_length = 35, ...)

Mean2default(x, digits = 2, n = length(x),
  mean.style = get_my_options()$apa.style$mittelwert$mean.style)

Median2default(x, digits = 2, n = length(x),
  median.style = get_my_options()$apa.style$mittelwert$median.style)

Prozent2default(x, digits = 0, n = length(x), exclude = NA,
  max_factor_length = 25)
}
\arguments{
\item{...}{weitere Objekte}

\item{data}{bei verwendung von Formeln}

\item{x}{Namen der  measure.vars}

\item{type}{mean, median}

\item{fun}{Function an plyr::ddply}

\item{fm}{formel}

\item{digits}{Nachkommastellen}

\item{measure.name, measure, by}{Variablen name des Ergebnisses. 
In Tabelle() measure.name  ="value"}

\item{.data}{Daten}

\item{na.rm}{NAs}

\item{conf.interval}{CIs}

\item{.drop}{anplyr::ddply}

\item{ci}{Grenzen der Konfidenzintervalle}

\item{x}{Objekt Vector oder auch Formel}

\item{na.rm}{Fehlende Werte}

\item{digits}{Dezimalstellen bei zB Mean2}

\item{...}{Weitere Argumente}
}
\value{
ein dataframe Objekt oder ein Character-String

Vector
}
\description{
Die Lagemasse werden ueber die Standard-Funktionen berechnet unterschied ist nur dass
Faktoren zu Zahlen transformiert werden und das \code{na.rm=TRUE} gesetzt wird.
CI = Hmisc::smean.cl.normal

Prozent2 wird in APA.formula errate_statistik2 verwendet
freq = Prozent2(x_NA, exclude, digits=digits.percent, rtrn = "df", max_factor_length = max_factor_length)
}
\examples{

#  Tabelle( hyper, chol0,chol1,chol6,chol12,by=~ g)

res<-stp25APA2:::berechne.default(hyper, 
                                  Cs(chol0,chol1,chol6,chol12), 
                                  by=~med+g, 
                                  measure="mean")
names(res)
 # res
 
berechne(hyper, "chol0" )
names(hyper)
hyper \%>\% berechne(chol0,chol1,chol6,chol12, by=~med+g)
mean2(rnorm(100))
sd2(rnorm(100))
CI(rnorm(100))
StdErr(rnorm(100))
mean2(rnorm(100))
set_my_options(prozent = list(digits = 1,
style = 2,
null_percent_sign= "."))

Prozent2default(factor(c(1,2,3,3,3,5), 1:5))

}
