% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intern_errate_statistik.R
\name{errate_statistik2}
\alias{errate_statistik2}
\title{Errate korekte Auswertung
Extrahieren der Variablen aus Formula.}
\usage{
errate_statistik2(Formula, data, caption = "caption", note = "note",
  type = NULL, na.action = na.pass, exclude = NA, include.n = TRUE,
  include.all.n = NULL, include.header.n = TRUE, include.total = FALSE,
  include.test = FALSE, include.p = TRUE, include.sig.star = FALSE,
  corr_test = "pearson", cor_diagonale_up = TRUE, max_factor_length = 35,
  order = FALSE, decreasing = FALSE, useconTest = FALSE,
  normality.test = FALSE, digits.mean = NULL, digits.percent = NULL,
  test_name = "Hmisc", ...)
}
\arguments{
\item{Formula}{Formel a~b}

\item{data}{data.frame}

\item{type}{type=c("auto", "freq", "mean", "median")}

\item{exclude}{fuer factor}

\item{include.all.n}{N ausgeben oder nicht}

\item{include.header.n}{Kopfzeile}

\item{include.test}{sig test bei auto fuer SPSS-like test=="wilcox.test"  test=="kruskal.test"}

\item{include.p}{p-Werte ja nein}

\item{include.sig.star}{logical Formatierung}

\item{corr_test}{default ist  "pearson"}

\item{cor_diagonale_up}{logical}

\item{max_factor_length}{interne Funktion}

\item{order}{logical interne Funktion}

\item{decreasing}{interne Funktion}

\item{useconTest}{logical auswahl des Sig.test   bis jetzt nur con_tests implementiert (t-test usw)}

\item{normality.test}{logical}

\item{test_name}{Welche Teste}

\item{...}{weitere Parameter}

\item{total}{logical   Hinzufuegen der Gesamtwerte bei Gruppenvergleichen}
}
\description{
Errate korekte Auswertung
Extrahieren der Variablen aus Formula.
}
