% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_etaSquared2.R
\name{etaSquared2}
\alias{etaSquared2}
\title{Eta-Quadrat}
\usage{
etaSquared2(x, type = 2, anova = FALSE, ...)
}
\arguments{
\item{x}{fit lm oder anova}

\item{type}{Anova Type default ist 2}

\item{anova}{Ausgabe der ANOVA Tabelle}

\item{...}{nicht benutzt}
}
\value{
Vector
}
\description{
Kopie von lsr::etaSquared
}
\examples{
fit1<-lm(y1~x1, anscombe)
#etaSquared2(aov (y1~x1, anscombe), anova=TRUE)
#etaSquared2(fit1, anova=TRUE)
etaSquared2(fit1 )

}
