% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-extract.R
\name{extract_param}
\alias{extract_param}
\title{extract_param}
\usage{
extract_param(x, ...)
}
\arguments{
\item{x}{Objekt}

\item{...}{weitere params}

\item{include.b, include.beta, include.ci, include.se, include.odds, include.odds.ci, include.statistic, include.p, include.stars, include.df, include.effects, conf.level, conf.method}{Parameter wie in APA_Tabelle beschrieben.}
}
\value{
data.frame tibble  "term","estimate", "beta","std.error" ,"statistic","p.value","stars","conf.high","conf.low","odds","odds.conf.low","odds.conf.high","df","group"
}
\description{
Summary fuer Regressionsmodelle. Kopie von broom::tidy
}
\examples{
fit <- lm(Sepal.Width ~ Petal.Length + Petal.Width, iris)
broom::tidy(fit)
broom::glance(fit)

summary(fit)
Ordnen(fit)
extract_param(fit)
APA_Table(fit)
APA_Table(fit, type="long")


}
