% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_.R
\name{Format2}
\alias{Format2}
\alias{Format2.matrix}
\alias{Format2.data.frame}
\alias{Format2.list}
\alias{Format2.default}
\alias{ffpvalue}
\alias{ffsigstars}
\alias{ffprozent}
\alias{ffprozent.default}
\alias{ffprozent.data.frame}
\alias{ffprozent.matrix}
\alias{ffmean}
\alias{ffmean_long}
\alias{ffmedian_long}
\alias{ffmedian}
\alias{ffCI}
\alias{ffreta}
\alias{fftest}
\title{Format2}
\usage{
Format2(x, ...)

\method{Format2}{matrix}(x, digits, ...)

\method{Format2}{data.frame}(x, digits = NULL, lead.zero = TRUE,
  type = "digits", scientific = FALSE, ...)

\method{Format2}{list}(x, digits = NULL, lead.zero = TRUE,
  type = "digits", scientific = FALSE, ...)

\method{Format2}{default}(x, digits = NULL, lead.zero = TRUE,
  OutDec = getOption("OutDec"), type = "digits", scientific = FALSE,
  nsmall = ifelse(is.null(digits), 0L, digits), ...)

ffpvalue(x, digits = options()$stp25$apa.style$p$digits,
  lead.zero = options()$stp25$apa.style$p$lead.zero,
  with.stars = options()$stp25$apa.style$p$with.stars, pval = x,
  lim_sig = ifelse(options()$prompt[1] == "HTML> ", "&lt;", "<"), ...)

ffsigstars(x, stars.value = options()$stp25$apa.style$p$stars.value,
  stars.symbols = options()$stp25$apa.style$p$stars.symbols)

ffprozent(prz, frq = NULL, ...)

\method{ffprozent}{default}(prz, frq = NULL,
  digits = options()$stp25$apa.style$prozent$digits,
  lead.zero = options()$stp25$apa.style$prozent$lead.zero,
  style = options()$stp25$apa.style$prozent$style,
  null_percent_sign = options()$stp25$apa.style$prozent$null_percent_sign,
  ...)

\method{ffprozent}{data.frame}(prz, frq = NULL, ...)

\method{ffprozent}{matrix}(prz, frq = NULL, ...)

ffmean(m, s = NULL, digits = rep_len(countDigits(m), 2),
  plusmin_sign = options()$stp25$apa.style$mittelwert$plusmin_sign,
  sym = options()$stp25$apa.style$mittelwert$plusmin_str, ...)

ffmean_long(m, s, mn, mx, digits = countDigits(m), ...)

ffmedian_long(m, iq, mn, mx, digits = countDigits(m), ...)

ffmedian(quant, digits = NA, style = options()$stp25$apa.style$median,
  sep = options()$stp25$apa.style$sep_element, ...)

ffCI(CIs, digits = 2,
  lead.zero = options()$stp25$apa.style$mittelwert$lead.zero,
  sep = options()$stp25$apa.style$sep_element,
  sep_1 = options()$stp25$apa.style$brackets[1],
  sep_2 = options()$stp25$apa.style$brackets[2], ...)

ffreta(x, digits = options()$stp25$apa.style$r$digits,
  lead.zero = options()$stp25$apa.style$r$lead.zero)

fftest(x, digits = options()$stp25$apa.style$Fstat$digits,
  lead.zero = options()$stp25$apa.style$Fstat$lead.zero)
}
\arguments{
\item{x}{vektor liste oder matrix}

\item{...}{alles an format}

\item{digits}{bei data.frame auch mehere Werte}

\item{lead.zero}{= TRUE}

\item{type}{= "digits"oder"signif}

\item{scientific}{= FALSE}

\item{OutDec}{= getOption("OutDec")}

\item{nsmall}{nicht aendern werden digite}

\item{pattern_pval}{nicht aendern P-Werte Namen}

\item{pattern_est}{nicht aendern Parameter Namen}

\item{pattern_df}{nicht aendern DF Namen}

\item{apa.style}{nicht aendern verwendeter Style als liste}

\item{col_names}{fuer fix_colnames col_names=NULL}

\item{translate}{fuer fix_colnames(x, translate= TRUE)}
}
\value{
Objekt als Character
}
\description{
Formatiere Zahlen zu Character f?r den Output \link{format}

Format2.matrix kann über digits
verschiedene nachkommastellen runden

Format2.data.frame
}
\examples{
##library(stp5)
# Projekt(OutDec = ",")
# options()
x<- rnorm(10)
df <- data.frame(Item=c("a", "b"),
                 x=x[1:2],
                 x2=c(1.2,2.3),
                 beta=  c(.22,.13),
                 x3=c(.42,.03),
                 p.value=c(0.02456,0.0398))

mx<- as.matrix(df[,-1])
#Projekt("html", OutDec = ".")
#Output(Format2(df, digits=c(1,2,2,2,2,3)))
#Format2(mx, digits=c(4,2,2,2,2,3))
Format2(x, digits=2, FALSE)
Format2(x, digits=3)
x<- list(a=1:5, b=rnorm(10))
length(x)
Format2(x, 2, FALSE)

#End()
x<-matrix(rnorm(10), ncol=2)
Format2.matrix(x[,1],c(1:5))
#  str(Format2(data.frame(a1=1:10/100,a2=1:10, a3=1:10)))
}
