% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA2_Reliability.R
\name{Index}
\alias{Index}
\title{Summen Index}
\usage{
Index(x, revcoded = FALSE, fun = "mean", na.rm = TRUE, digits = 4,
  max.level = NA, min.level = NA, output = "index", ...)
}
\arguments{
\item{x}{Data.frame}

\item{revcoded}{position der umzukodierenden Items}

\item{fun}{mean oder sum}

\item{na.rm}{Fehlende Werte}

\item{digits}{default   4}

\item{max.level}{maximaler Mevel}
}
\value{
Vektor
}
\description{
Eigendlich nur eine Summenfunktion mit der Erweiterung zum Umcodieren
}
