% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{intern_cor}
\alias{intern_cor}
\alias{corr_tabel}
\alias{corr_tabel.data.frame}
\alias{corr_tabel.formula}
\alias{corr_tabel2}
\alias{corr_1}
\alias{corr_2}
\alias{Corr2}
\alias{Corr1}
\title{interne Funktion fuer die Korrelationen}
\usage{
corr_tabel(x, ...)

\method{corr_tabel}{data.frame}(.data, ..., by = NULL, groups = NULL,
  type = c("pearson", "spearman"), cor_diagonale_up = TRUE,
  sig.star = TRUE, p.value = FALSE, include.mean = FALSE,
  include.n = FALSE)

\method{corr_tabel}{formula}(Formula, data, groups = NULL,
  type = c("pearson", "spearman"), exclude = NA, subset,
  na.action = na.pass, cor_diagonale_up = TRUE, sig.star = TRUE,
  p.value = FALSE, include.mean = FALSE, include.n = TRUE)

corr_tabel2(..., cor_diagonale_up = TRUE, sig.star = TRUE,
  p.value = FALSE, include.mean = FALSE, include.n = TRUE)

corr_1(x, type = "pearson")

corr_2(x, y = NULL, type = "pearson")

Corr2(x, y = NULL, type = "pearson", sig.star, ...)

Corr1(y, n = nrow(y), type = "pearson", include.p = FALSE,
  include.sig.star = TRUE, cor_diagonale_up = TRUE, dimnames = FALSE)
}
\arguments{
\item{.data}{ist das erste Obkjekt}

\item{by}{wenn das erste Argument die Daten sind kann hier die
Gruppe gewaelt werdendefault = NULL}

\item{groups}{groups spezielle Gruppen  wie zB Korrelation zwischen den Geschlechtern al}

\item{type}{Cor-Typ   default = "pearson" , type=c("pearson","spearman")}

\item{cor_diagonale_up}{diagonale   Linksbuendig = TRUE}

\item{sig.star, p.value}{Sternchen oder p-Werte}

\item{include.mean}{Mittelwert mit SD}

\item{include.n}{Anzahl N}

\item{Formula}{Ist das erste Element}

\item{data}{daten fuer Formula-Methode}

\item{exclude, subset, na.action}{fuer Formula Data (geht aber nicht)}
}
\value{
corr_tabel: liste mit "r",        "n",        "P",        "mean",     "row_name"

corr_tabel2: data.frame()
}
\description{
Die Funktion \code{corr_tabel}: ist die interne Funktion die die Berechnungen erstellt.

corr_tabel.data.frame ist die Data.Frame- methode bei der die Variablennamen ueber das ...
Argument uebergeben werden. Die Funktion arbeitet intern mit der corr_tabel.formula Funktion.

corr_tabel2 ist die Ausgabe von corr_tabel als Formatierte Tabelle

Die Interne Fznction \code{Corr2} wird in APA2 verwendete Korrelation

Die Interne Fznction \code{Corr2} wird in APA2 verwendete Korrelation
}
\examples{

#-- corr_tabel.data.frame
corr_tabel(data, a, b, d)

#-- corr_tabel.formula

corr_tabel(a+b~d+c, data)
corr_tabel(a+b+c ~ d, data)

corr_tabel(a+b+c~d, data, groups = ~g)

#--corr_tabel2

corr_tabel2(data, a, b, d)

#-- corr_1
stp25APA2:::corr_1(data[1:3])
#-- corr_2
# stp25APA2:::corr_2(data[1:3], data[4])


#-- Corr2
# APA2(a+b+c~d, data )
stp25APA2:::Corr2(data[1:3], data[4], "pearson", TRUE)


#-- Corr1
# APA2(~a+b+cd, data)
stp25APA2:::Corr1(data[1:3], dimnames=TRUE)
stp25APA2:::Corr1(data[1:3], dimnames=TRUE, include.p=TRUE)
}
