% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kano.R
\name{Kano}
\alias{Kano}
\alias{APA2.Kano}
\title{Analyze Kano type items.}
\usage{
Kano(X, grouping = NULL, type = 5, umcodieren = FALSE, rm_Q = 10000,
  rm_I = 10000, methode = 1, vars_func = NULL, vars_dysfunc = NULL,
  data = grouping, subset = NULL, na.action = NULL, ...)

\method{APA2}{Kano}(x, caption = "", note = "note", formula = x$formula,
  data = x$molten, digits = options()$stp4$apa.style$prozent$digits[1],
  prop.table = TRUE, ...)
}
\arguments{
\item{X}{data.frame mit parweisen Kano Fragen}

\item{grouping}{Dataframe mit Gruppen}

\item{type}{Fragetype entwerer vollstaendig (5) oder gekuerzt (3)}

\item{umcodieren}{logical False}

\item{rm_Q}{Remove Q Kategorien Q enfernen Anzahl an erlaubten Qs}

\item{rm_I}{Remove I Kategorien I enfernen Anzahl an erlaubten Is}

\item{methode}{eie sind die Items geornet}

\item{vars_func}{Welche Items sind die Funktionalen}

\item{vars_dysfunc}{Welche Items sind die Dys-Funktionalen}

\item{...}{\itemize{
   \item \code{type=c(1,2)}  Streudiagram (1) oder Balken (2)
   \item \code{groups= NULL} wenn ein Dataframe uebergeben dan sind das die Gruppen
   \item \code{legend.position = list(x="right", y=NULL) } - Beim Streudiagram die
   Legende bei Barplot \code{scale=list()}
   \item \code{txt.bg=list(m="M",i= "I", o="O", a="A")} Beim Streudiagram der
   Hintergrundtext
   \item \code{jitter=TRUE} Beim Streudiagram Verrauschen der Items (wegen besserer
   lesbarkeit)
   \item \code{col.bg="gray95"}
   \item \code{col = NA}
   \item \code{cex=0.70}
   \item \code{cex.items=cex*1}

   \item \code{cex.lab=cex*1.07}
   \item \code{cex.legend=cex*1.1}
   \item \code{ylab="Zufriedenheitsstiftung (CS+)"}
   \item \code{xlab="Unzufriedenheitsstiftung (CS-)"}
   \item \code{CS.minus = "CS minus"}
   \item \code{center.axis=FALSE}
   \item \code{use.labels=FALSE  Zahlen oder Text als Beschriftung bei den Streudiagrammen}
   \item \code{use.total.strength =TRUE  groese der Schrift als Indikator fuer Total Strenght}
   \item \code{use.categorie=TRUE  M, O, A, oder i mit in die Labels ausgeben }

}}
}
\value{
Liste mit
value: data.frame mit der Kano-Kodierung
scors: Scors,
data= data,
molten=molten,
formula= fm,
removed=Errorrs,
N=nrow(data),
Attributes= c("Must-be","One-dimensional", "Attractive","Indifferent","Reverse", "Questionable"),
answers
}
\description{
Transformiert Kano-Fragebogen zu Kano-Kodierung
http://www.eric-klopp.de/texte/angewandte-psychologie/18-die-kano-methode

M O A I R Q Heufigkeit

max Category

M>O>A>I  max Category mit Hirarchie M Wichtiger als O usw.
also wen der Unterschied zwischen den zwei am hoechsten gelisteten Atributen
zwei Kategorien gleich ist,  5% Schwelle, dann gilt die Regel M>O>A>I

Total Strength als zweite Masszahl gibt an wie hoch der Anteil an bedeutenden
Produktmerkmalen ist.

Category Strength ist eine Masszahl die die angibt ob eine Anforderung nur in
eine Kategorie gehoert

CS plus	Indx Positiv  CS.plus=  (A+O)/(A+O+M+I)

CS minus	Index Negativ CS.minus= (O+M)/(A+O+M+I)

Chi-Test	Eigendlich unsinn Testet ob Verteilung von M, A, O und I gleich ist

Fong-Test Vergleich der zwei Haeufigsten-Kategorien gegenueber der Gesammtzahl
Ergebnis ist entweder ein signifikente oder ein nicht signifikente Verteilung.


 \subsection{Basis-Faktoren (Mussfaktoren)}{
 Basis-Merkmale (M...Mustbe) werden vom Kunden Vorausgesetzt schaffen
 unzufriedenheit wenn sie nicht vorhanden sind.
 }
  \subsection{Leistungs- Faktoren}{
 Leistungs-Merkmale (O...One-dimensional) werden vom Kunden verlangt
 }
 \subsection{Begeisterung-Faktoren}{
 Begeisterungs-Merkmale (A...Attractive) Kunde rechnet nicht damit hebt das
 Produkt vom Konkurenten ab.
 }
 \subsection{Unerhebliche- Faktoren }{
 Unerhebliche-Merkmale (I...Indifferent) werden vom Kunden Ignoriert.
 }
}
