% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA2_Correlation.R
\name{Korrelation}
\alias{Korrelation}
\alias{APA_Correlation}
\alias{corr_tabel}
\alias{corr_tabel.data.frame}
\alias{corr_tabel.formula}
\alias{corr_tabel2}
\alias{corr_1}
\alias{corr_2}
\alias{Corr2}
\alias{Corr1}
\title{Korrelationstabelle}
\usage{
APA_Correlation(x, ..., caption = "Korrelation", note = NULL)

corr_tabel(x, ...)

\method{corr_tabel}{data.frame}(.data, ..., by = NULL, groups = NULL,
  type = c("pearson", "spearman"), cor_diagonale_up = TRUE,
  sig.star = TRUE, p.value = FALSE, include.mean = FALSE,
  include.n = FALSE)

\method{corr_tabel}{formula}(Formula, data, groups = NULL,
  type = c("pearson", "spearman"), exclude = NA, subset,
  na.action = na.pass, cor_diagonale_up = TRUE, sig.star = TRUE,
  p.value = FALSE, include.mean = FALSE, include.n = TRUE)

corr_tabel2(..., cor_diagonale_up = TRUE, sig.star = TRUE,
  p.value = FALSE, include.mean = FALSE, include.n = TRUE)

corr_1(x, type = "pearson")

corr_2(x, y = NULL, type = "pearson")

Corr2(x, y = NULL, type = "pearson", sig.star, ...)

Corr1(y, n = nrow(y), type = "pearson", include.p = FALSE,
  include.sig.star = TRUE, cor_diagonale_up = TRUE, dimnames = FALSE)
}
\arguments{
\item{x}{Objekt data.frame oder formula}

\item{...}{alles weiter zB namen der Variablen}

\item{caption, note}{Ueberschrift  Fussnote Korrelation}

\item{.data}{ist das erste Obkjekt}

\item{by}{wenn das erste Argument die Daten sind kann hier die
Gruppe gewaelt werdendefault = NULL}

\item{groups}{groups spezielle Gruppen  wie zB Korrelation zwischen den Geschlechtern al}

\item{type}{Cor-Typ   default = "pearson" , type=c("pearson","spearman")}

\item{cor_diagonale_up}{diagonale   Linksbuendig = TRUE}

\item{sig.star, p.value}{Sternchen oder p-Werte}

\item{include.mean}{Mittelwert mit SD}

\item{include.n}{Anzahl N}

\item{Formula}{Ist das erste Element}

\item{data}{daten fuer Formula-Methode}

\item{exclude, subset, na.action}{fuer Formula Data (geht aber nicht)}
}
\value{
corr_tabel: liste mit "r",        "n",        "P",        "mean",     "row_name"

corr_tabel2: data.frame()
}
\description{
Diese Funktion erstellt Korrelationstabelle (Interkorrelationen mit hilfe
der Funktion \code{\link[Hmisc]{rcorr}}. Erlaubt ist die getrennte Auswertung ueber groups bzw mit by
siehe auch Beispiel unten.
Die Funktion arbeitet mit der Pipe oder auch mit Formeln.

APA_Correlation arbeitet intern mit ser Funktion corr_tabel.

corr_tabel ist die interne Funktion die die Berechnungen erstellt

corr_tabel.data.frame ist die Data.Frame- methode bei der die Variablennamen ueber das ...
Argument uebergeben werden. Die Funktion arbeitet intern mit der corr_tabel.formula Funktion.

corr_tabel2 ist die Ausgabe von corr_tabel als Formatierte Tabelle

Die Interne Fznction \code{Corr2} wird in APA2 verwendete Korrelation

Die Interne Fznction \code{Corr2} wird in APA2 verwendete Korrelation
}
\examples{
n<- 2*8
e<- rnorm(n)*10
data<- data.frame(a=rnorm(n) + e,
                  b=rnorm(n), c=rnorm(n), d=rnorm(n) + e,
                  g=gl(2, 8, labels = c("Control", "Treat")))
#
APA_Correlation(~a+b+c, data, caption="~a+b+c")

APA_Correlation(~a+b+c, data, caption="mit Output: ~a+b+c")


 # library(arm)
 # windows(7,7)
 # corrplot(data[Cs(a,b,c,d)], abs=TRUE, n.col.legend=7)

 # install.packages("PerformanceAnalytics")

 #library("PerformanceAnalytics")
 #?chart.Correlation#(decathlon2.active[, 1:6], histogram=TRUE, pch=19)
 #data(managers)
 #chart.Correlation(managers[,1:8], histogram=TRUE, pch="+")

#-- corr_tabel.data.frame
corr_tabel(data, a, b, d)

#-- corr_tabel.formula

corr_tabel(a+b~d+c, data)
corr_tabel(a+b+c ~ d, data)

corr_tabel(a+b+c~d, data, groups = ~g)

#--corr_tabel2

corr_tabel2(data, a, b, d)

#-- corr_1
stp25APA2:::corr_1(data[1:3])
#-- corr_2
# stp25APA2:::corr_2(data[1:3], data[4])


#-- Corr2
# APA2(a+b+c~d, data )
stp25APA2:::Corr2(data[1:3], data[4], "pearson", TRUE)


#-- Corr1
# APA2(~a+b+cd, data)
stp25APA2:::Corr1(data[1:3], dimnames=TRUE)
stp25APA2:::Corr1(data[1:3], dimnames=TRUE, include.p=TRUE)
}
