% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA2_xtabs.R
\name{Kreuztabellen}
\alias{Kreuztabellen}
\alias{APA_Xtabs}
\alias{APA2.loglm}
\title{Kreuztabellen}
\usage{
APA_Xtabs(x, caption = "", note = "", digits = NULL, test = FALSE,
  type = c("fischer", "odds", "sensitivity", "chisquare", "correlation", "r"),
  include.total = FALSE, include.total.columns = include.total,
  include.total.sub = include.total, include.total.rows = include.total,
  include.percent = TRUE, margin = NULL, add.margins = seq_along(dim(x)),
  ...)

\method{APA2}{loglm}(x, ...)
}
\arguments{
\item{x}{xtabs Objekt geht aber auch mit table aber sollte wegen der Beschriftungen (labels) nicht verwendet werden.}

\item{caption}{Ueberschrift}

\item{note}{Fussnote}

\item{digits}{fuer Sensitivitaet sonst ueber Start() einstellen}

\item{test}{Signifikanztest TRUE/FALSE  type = FALSE}

\item{type}{type = c("fischer", "odds","sensitivity", "chisquare","correlation", "r")bei 2x2-Tabellen Fischer-test und}

\item{include.total}{default = FALSE total/sum}

\item{include.total.columns}{columns total/sum}

\item{include.total.sub}{sum total/sum}

\item{include.total.rows}{rows total/sum}

\item{include.percent}{Prozent TRUE/FALSE}

\item{margin}{Zeilenprozent, Spaltenprozent}

\item{...}{weitere Argumente}
}
\value{
HTML oder Text
}
\description{
Formatiert xtabs() 2x2 Tabellen werden mit Haufigkeit wahlweise mit Prozent
(verhalten wird ueber \code{margin = 2}) geseuert. Berechnet werden mittels
 \link{fisher.test} die Odds-Ratio mit 95%CI und die p-Werte. Alternatibe ist die Funktion
 \link{oddsratio} aus dem Paket \code{vcd}. Weiters laesst sich mit der Option \code{type = 2}
 eine sensitivit?tsanalyse erstellen.
 Bei NxM-Tabellen wird als Test-Statistik Pearson und der
 Kontingentkoeffizient berechnet alternativ steht auch  der Phi-Coefficient zur verf?gung
 auch hier mit \code{type = 2}. die Berechnung erfolgt hier mit \link{assocstats}
 aus dem Packet \code{vcd}.
}
\examples{
# library(stp25data)
# Projekt("html")
hkarz$LAI<- factor(hkarz$lai, 0:1, c("pos", "neg"))
hkarz$Tzell<- cut(hkarz$tzell, 3, c("low", "med", "hig"))


xtab <- xtabs(~ gruppe+LAI, hkarz)
APA2(xtab, caption="Harnblasenkarzinom", test=FALSE)
APA2(xtab, type="sens", test=TRUE, caption = "type=sens")
APA2(xtab, type="sens", caption = "geht nur mit teat=TRUE +  type=sens")
APA2(xtabs(~ gruppe+Tzell, hkarz), caption="APA_Xtabs: 2x3 Tabelle", test=FALSE)
APA2(xtabs(~ gruppe+LAI+Tzell, hkarz), caption="APA_Xtabs: 2x2x3 Tabelle", test=FALSE)


APA2(xtab, include.total.columns=TRUE, caption = "include.total.columns")
APA2(xtab, include.total.sub=TRUE, caption = "include.total.sub")




xtab <- xtabs(~ gruppe+Tzell, hkarz)
APA2(xtab, test=FALSE, caption="APA2: 2x3 Tabelle")
APA_Xtabs(xtab, caption="APA_Xtabs: 2x3 Tabelle")

# Fit Log-Linear Models by Iterative Proportional Scaling
library(MASS)

fit<-loglm(~ Type + Origin, xtabs(~ Type + Origin, Cars93))
APA2(fit)
}
