% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-mean2.R
\name{Prozent}
\alias{Prozent}
\title{Prozent Berechnen}
\usage{
Prozent(x, digits = 1, continuous = 3, breaks = NULL,
  labels = NULL, count_factor = c("yes", "ja", "T", "TRUE", 1),
  retur_tabel = FALSE)
}
\arguments{
\item{x}{Objekt Vector oder auch Formel}

\item{digits}{Dezimalstellen bei zB Mean2}

\item{continuous}{fuer cut}

\item{breaks}{fuer cut}

\item{labels}{fuer cut}

\item{count_factor}{wiess nicht}

\item{retur_tabel}{wiess nicht}

\item{exclude}{Fehlende Werte}

\item{max_factor_length}{wie viele Faktoren}

\item{continuous}{fuer cut}

\item{...}{Weitere Argumente}
}
\value{
Vector
}
\description{
Prozent fuer
}
\examples{
mean2(rnorm(100))
sd2(rnorm(100))
CI(rnorm(100))
StdErr(rnorm(100))
mean2(rnorm(100))
}
