% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_RMSE.R, R/Helper_etaSquared2.R
\name{RMSE}
\alias{RMSE}
\alias{RMSE.default}
\alias{RMSE.mlm}
\alias{RMSE}
\alias{RMSE.default}
\alias{RMSE.mlm}
\title{RMSE}
\usage{
RMSE(x, ...)

\method{RMSE}{default}(x, ...)

\method{RMSE}{mlm}(x, ...)

RMSE(x, ...)

\method{RMSE}{default}(x, ...)

\method{RMSE}{mlm}(x, ...)
}
\arguments{
\item{x}{fit-Objekt lm glm}

\item{...}{weitere Objekte nicht benutzt}

\item{x}{fit-Objekt lm glm}

\item{...}{weitere Objekte nicht benutzt}
}
\value{
Data.frame mit Namen

Data.frame mit Namen
}
\description{
## The RMSE is the square root of the variance of the residuals.
Compute the root mean squared error
(see \code{\link{sigma}})

## The RMSE is the square root of the variance of the residuals.
Compute the root mean squared error
(see \code{\link{sigma}})
}
\examples{
#library(haven)


#library(haven)


}
