% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndr.R
\name{rndr_}
\alias{rndr_}
\alias{rndr_median}
\alias{rndr_mean}
\alias{rndr_ods}
\alias{rndr_percent}
\alias{rndr_P}
\alias{rndr_Stars}
\alias{rndr_corr}
\alias{rndr_r}
\alias{rndr_df}
\alias{rndr_CI}
\alias{rndr_F}
\alias{rndr_T}
\alias{rndr_H}
\alias{rndr_W}
\alias{rndr_W}
\alias{rndr_lm}
\alias{rndr_X}
\alias{rndr_Chisq}
\alias{rndr_fischer}
\alias{rndr_gfi_cfa}
\alias{rndr_agfi_cfa}
\alias{rndr_rmsea_cfa}
\alias{rndr_Chisq_cfa}
\alias{rndr_rmsea_cfa}
\alias{rndr_cfi_cfa}
\alias{rndr_nfi_cfa}
\title{render for html-output}
\usage{
rndr_(...)

rndr_median(m, iqr, digits)

rndr_mean(m, s, digits)

rndr_ods(x, digits = 2)

rndr_percent(prozent, anzahl, percent = TRUE,
  digits = options()$stp25$apa.style$prozent$digits[1],
  percentage_str = options()$stp25$apa.style$prozent$percentage_str,
  style = options()$stp25$apa.style$prozent$style,
  null_percent_sign = options()$stp25$apa.style$prozent$null_percent_sign)

rndr_P(p_val, include.symbol = TRUE, include.bracket = FALSE)

rndr_Stars(p_val)

rndr_corr(r, p, df)

rndr_r(r, include.symbol = TRUE)

rndr_df(df1, df2 = NULL)

rndr_CI(x, digits = 2, sep = options()$stp25$apa.style$sep_element,
  sep_1 = options()$stp25$apa.style$brackets[1],
  sep_2 = options()$stp25$apa.style$brackets[2])

rndr_F(F_val, df1, df2, p = NULL)

rndr_T(F_val, df1, p = NULL)

rndr_H(F_val, df1, p = NULL)

rndr_W(F_val, p = NULL)

rndr_W(F_val, p = NULL)

rndr_lm(F_val, df1, df2, p, r2, ar2)

rndr_X(x, df1, df2 = NULL, p = NULL)

rndr_Chisq(x, df, p)

rndr_fischer(or, p)

rndr_gfi_cfa(gfi)

rndr_agfi_cfa(agfi)

rndr_rmsea_cfa(rmsea)

rndr_Chisq_cfa(chsq, df = 1)

rndr_rmsea_cfa(rmsea)

rndr_cfi_cfa(nfi)

rndr_nfi_cfa(nfi)
}
\arguments{
\item{...}{alles an format}
}
\value{
Objekt als Character
}
\description{
Formatiere Zahlen zu Character feur die Ausgabe. Im gegensatz zu Format2()
und ff() werden die Zahlen mit zB p=.002 ausgegeben.

median

mean

Odds

rndr_percent()
Percentages are also most clearly displayed in parentheses with no decimal places:
Nearly half (49%) of the sample was married.

Prozent

APA Textausgabe
p-Werte \code{rndr_P()}

rndr_corr
Correlations are reported with the degrees of freedom (which is N-2)
in parentheses and the significance level:
The two variables were strongly correlated, r(55) = .49, p < .01.

CI-Wert\code{rndr_CI()}  rndr_CI(matrix(c(NA, 1:10, NA), ncol=2))

F-Wert\code{rndr_F()}

T-Wert\code{rndr_T()}
T Tests are reported like chi-squares, but only the degrees of freedom are
 in parentheses. Following that, report the t statistic (rounded to two decimal places)
  and the significance level.
 There was a significant effect for gender, t(54) = 5.43, p < .001, with men receiving higher scores than women.

H-Wert\code{rndr_H()}

W-Wert\code{rndr_W()}

U-Wert\code{rndr_U()}

Lm-Wert\code{rndr_ lm()}

Chi-Wert\code{rndr_X()}

Fischer-test\code{rndr_fischer()}

CFA Confirmatorische Faktoranalyse

Backhaus Multivariate Analysemethoden 11 AuflageSeite 383
GIF Goodness-of-Fit-Index >=.9

AGIF Adjusted-Goodness-of-Fit-Index

SRMR

Moosbrugger, Kelava 2012 Testtheorie 2. Auflage Seite 339
CHISQ Chi-Quadrat/df 0 ,2, 3

RMSEA Root-Mean-Square-Error of Approximation 0, 0.050, 0.080

CFI Comparative-Fit-Index .970-1.00, .950-.969

NFI Normed-Fit-Index .950-1.00 , .900-.949
}
\details{
http://winvector.github.io/APAsig/APAsig.html
Format quality of a linear regerssion in
roughly in "APA Style" ( American Psychological Association ).
see: http://my.ilstu.edu/~jhkahn/apastats.html
     https://web2.uconn.edu/writingcenter/pdf/Reporting_Statistics.pdf
Not linking direclty to the style guide as they charge for it.
}
\examples{
rndr_(1.234, 3)
rndr_percent(c(.2568, 99, 0.02568) , c(4, 569, 25), digits = 1)
rndr_percent(10, 3, F, 2)

rndr_P(c(1,.25,.08,0.05,0.01,0.0001))
rndr_corr(-.548369,0.005896,55)
rndr_T(25.45, 45, .0045)
rndr_H(25.45, 45, .0045)
 rndr_gfi_cfa(c(1,.9,.89))
rndr_agfi_cfa(c(1,.9,.89))
rndr_rmsea_cfa(c(1,.9,.89))
rndr_Chisq_cfa(c(0,2,3,2.01,3.4))
rndr_rmsea_cfa(c(0, 0.050, 0.080, .051, .081) )
rndr_cfi_cfa(c(.970,1.00, .950-.969,.8))
rndr_nfi_cfa(c(.950, 1.00 , .900,  .949))
}
