% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stp25APA2.R
\docType{stp25APA2}
\name{stp25APA2}
\alias{stp25APA2}
\alias{stp25APA2-package}
\title{Convert Statistical Analysis Objects into APA-Tables}
\description{
Convert statistical analysis objects from R into APA-Tabls.
}
