% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA2_T_test.R
\name{t_test}
\alias{t_test}
\alias{APA_Ttest}
\alias{cohens.d}
\title{T-Test}
\usage{
APA_Ttest(x, data, var.equal = FALSE, paired = FALSE,
  alternative = "two.sided", include.mean = TRUE, include.d = TRUE,
  include.mean.diff = TRUE, include.test = TRUE, digits = 2, ...)

cohens.d(x, ...)
}
\arguments{
\item{x}{Formula}

\item{data}{Data.frame}

\item{var.equal, paired, alternative}{an t.Test var.equal = FALSE  "two.sided"}

\item{include.mean}{Mittelwerte mit SD}

\item{include.d}{Cohens d}

\item{include.mean.diff}{Mittlere Differenzen}

\item{digits}{nachkommastellen}

\item{...}{an Output}

\item{include.se}{Standardfehler}

\item{x, y}{formel oder x, y}

\item{...}{}
}
\value{
list

vector
}
\description{
Berechnung von Mittelwerten und Mittelwertdifferenzen

Computes Cohen's d
Cohen, J. (1988).  Statistical power analysis for the behavioral sciences (2nd ed.).  New York:Academic Press.
}
\examples{
 
#require(stpvers)

#Projekt("html")

APA_Ttest(m1+m2 ~ geschl, varana)
varanax<-Melt2(m1+m2~nr,varana , key="time", value="m")

# broom::tidy(with(varana, t.test(m1,m2 ) ))
# broom::tidy(  t.test(m~time, varanax, var.equal=FALSE)) )

APA_Ttest(m~time, varanax, paired = TRUE)

APA_Ttest(m~time, varanax, var.equal=TRUE)
APA_Ttest(m~time, varanax, var.equal=FALSE)
#broom::tidy(t.test(m~time, varanax, var.equal=TRUE))
#broom::tidy(t.test(m~time, varanax, var.equal=FALSE))

#End()

#require(stpvers)
set.seed(45)                        ## be reproducible 
 x <- rnorm(10, 10, 1)                
 y <- rnorm(10, 5, 5)

cohens.d(x, y)
varanax<-Melt2(m1+m2~nr,varana , key="time", value="m")
cohens.d(m~time, varanax )

}
