% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression1.R
\name{test_regression}
\alias{test_regression}
\title{Testing Linear Regression Models}
\usage{
test_regression(x, include.F = TRUE, include.heteroskedasticity = TRUE,
  include.durbin = TRUE, include.levene = FALSE, include.bartlett = FALSE,
  caption = "Testing Linear Regression Models")
}
\arguments{
\item{x}{model- fit}

\item{include.F}{Fsratistik}

\item{include.heteroskedasticity}{Breusch-Pagan test}

\item{include.durbin}{autocorrelation}

\item{include.levene}{homogeneity of variance across groupssiehe T-Test}

\item{include.bartlett}{Homogeneity of Variances siehe T-Test}

\item{caption}{}
}
\description{
Durbin-Watson test for autocorrelation of disturbances. Ist nur bei Zeitreihendaten sinnvoll.
Computes Levene's test for homogeneity of variance across groups.
Bartlett Test of Homogeneity of Variances
Breusch-Pagan test against heteroskedasticity.
variance inflation factor. VIF values over 5 are troubling, should probably investigate anything over 2.5.
}
