% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{VIF}
\alias{VIF}
\alias{VIF.default}
\alias{VIF2}
\alias{VIF.merModLmerTest}
\title{VIF Funktion}
\usage{
VIF(x, ...)

\method{VIF}{default}(fit, ...)

VIF2(..., caption = "VIF")

\method{VIF}{merModLmerTest}(fit, ...)
}
\description{
variance inflation factor, aka VIF  values over 5 are troubling.
should probably investigate anything over 2.5.
Quelle: https://hlplab.wordpress.com/2011/02/24/diagnosing-collinearity-in-lme4/
Kopie der \link{car::vif} funktion
}
\examples{

set.seed(0815)
fit<-lm(prestige ~ income + education, data=Duncan)
car::vif(fit)
VIF(fit)
}
\author{
Wolfgang Peter
}
