% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA2_Kaplan_Meier.R, R/APA_.R
\name{APA.survfit}
\alias{APA.survfit}
\alias{APA.survdiff}
\alias{APA}
\alias{APA.ScalarIndependenceTest}
\alias{APA.biVar}
\alias{APA.default}
\alias{APA.htest}
\alias{APA.lm}
\alias{APA.glm}
\alias{APA.formula}
\title{APA}
\usage{
\method{APA}{survfit}(fit, ...)

\method{APA}{survdiff}(fit)

APA(x, ...)

\method{APA}{ScalarIndependenceTest}(x, ...)

\method{APA}{biVar}(x, ...)

\method{APA}{default}(x, ...)

\method{APA}{htest}(x, ...)

\method{APA}{lm}(x, ...)

\method{APA}{glm}(x, ...)

\method{APA}{formula}(x, data, type = "mean", digits = 2, ...)
}
\arguments{
\item{...}{weitere Objekte nicht benutzt}

\item{x}{Objekt}
}
\value{
Character Vector mit einem oder meheren Eintraegen
}
\description{
APA Funktion wie \link{APA2} hier werden aber keine Tabellen oder HTML ausgegeben.
}
\examples{
APA(mpg ~ cyl, mtcars)
APA(coin::wilcox_test(mpg ~ factor(vs), mtcars))
APA(Hmisc::spearman2(mpg ~ factor(vs), mtcars))
APA(glm(vs ~ mpg, mtcars, family = binomial()))
APA(lm(mpg ~ drat + wt + qsec, mtcars))
APA(wilcox.test(mpg ~ vs, mtcars))
APA(t.test(mpg ~ vs, mtcars))

}
