% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APA.R, R/corr.R, R/default.R, R/regression1.R,
%   R/reliability.R, R/surv.R, R/t_test.R
\name{APA}
\alias{APA}
\alias{APA.NULL}
\alias{test_APA}
\alias{test_APA2}
\alias{APA.biVar}
\alias{APA.default}
\alias{APA.formula}
\alias{APA.lm}
\alias{APA.glm}
\alias{APA.stp25_relibility}
\alias{APA.survfit}
\alias{APA.survdiff}
\alias{APA.htest}
\alias{APA.ScalarIndependenceTest}
\title{APA Style Text-Ausgabe}
\usage{
APA(x, ...)

\method{APA}{NULL}(x, ...)

test_APA(x)

test_APA2(x)

\method{APA}{biVar}(x, ...)

\method{APA}{default}(x, ...)

\method{APA}{formula}(x, data, type = "mean", digits = 2, ...)

\method{APA}{lm}(x, include.r = TRUE)

\method{APA}{glm}(x, ...)

\method{APA}{stp25_relibility}(x)

\method{APA}{survfit}(fit, ...)

\method{APA}{survdiff}(fit)

\method{APA}{htest}(x, ...)

\method{APA}{ScalarIndependenceTest}(x, ...)
}
\arguments{
\item{x}{Objekt}

\item{...}{weitere Objekte nicht benutzt}

\item{include.r}{APA.lm: R-Squar}
}
\value{
Character Vector mit einem oder meheren Eintraegen
}
\description{
APA erstellt einen einzelnen  Text als Ergebnis.

APA.biVar: Korrelation mit Hmisc::spearman2

APA.htest: T-Test
}
\examples{

APA(mpg ~ cyl, mtcars)
APA(glm(vs ~ mpg, mtcars, family = binomial()))
APA(lm(mpg ~ drat + wt + qsec, mtcars))

  # Correlation
APA(Hmisc::spearman2(mpg ~ factor(vs), mtcars))


APA(Reliability(~F3+F2+F10+F11, fkv, check.keys =TRUE))
 # T-Test 
APA(coin::wilcox_test(mpg ~ factor(vs), mtcars))
APA(wilcox.test(mpg ~ vs, mtcars))
APA(t.test(mpg ~ vs, mtcars))
}
