% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kano.R
\name{Kano}
\alias{Kano}
\alias{Kano.data.frame}
\alias{Kano.formula}
\alias{Kano_Auswertung}
\alias{kano_plot2}
\title{Analyze Kano type items.}
\usage{
Kano(x, ...)

\method{Kano}{data.frame}(x, ...)

\method{Kano}{formula}(x, data, subset, na.action, ...)

Kano_Auswertung(x, caption = "", note = "", formula = x$formula,
  data = x$molten[-1], digits = options()$stp4$apa.style$prozent$digits[1],
  include.n = TRUE, include.percent = TRUE, include.total = TRUE,
  include.test = TRUE, rnd_output = TRUE, ...)

kano_plot2(x, main = "", xlim = c(0, 1), ylim = c(0, 1), mar = c(0, 1,
  2, 1), legend.position = list(x = "right", y = NULL), legend.title = NULL,
  my.lines = "circle", col.bg = "gray95", col = NA, txt.bg = list(m =
  "M", i = "I", o = "O", a = "A"), cex.bg = 12, jitter = TRUE, cex = 1,
  cex.items = cex * 1, cex.lab = cex * 1.07, cex.legend = cex * 1.1,
  ylab = "Zufriedenheitsstiftung (CS+)",
  xlab = "Unzufriedenheitsstiftung (CS-)", center.axis = FALSE,
  use.labels = TRUE, use.total.strength = TRUE, use.categorie = TRUE,
  use.points = FALSE, ...)
}
\arguments{
\item{x, data}{Daten mit/oder formula}

\item{...}{}

\item{caption, note}{Ueberschrift}

\item{formula, data}{intern daten aus Kano-Objekt}

\item{digits}{Nachkommastellen}

\item{include.n}{Anzahl}

\item{include.percent}{Prozent}

\item{include.total}{N und Total}

\item{include.test}{Fong und Chie-Test}

\item{rnd_output}{Intern fuer Plot bei FALSE ausgabe als Zahl}

\item{main}{Header}

\item{mar, xlim, ylim}{Plot Limits}

\item{legend.position, legend.title}{legend.position = list(x="right", y=NULL) 
Beim Streudiagram die Legende bei Barplot  scale=list}

\item{my.lines}{Grafik Element Halbkreis fuer Indiferent}

\item{col}{Farbe der Punkte}

\item{txt.bg, cex.bg, col.bg}{Hintergrund mit den Grossen Bchstaben}

\item{jitter}{Rauschen bei Uberlappung}

\item{cex, cex.items, cex.lab, cex.legend}{Schriftgroesse}

\item{ylab, xlab}{x-y Beschrifung}

\item{center.axis}{center.axis=FALSE}

\item{use.labels}{use.labels=FALSE  Zahlen oder Text als 
Beschriftung bei den Streudiagrammen}

\item{use.total.strength}{use.total.strength=TRUE  groese der Schrift 
als Indikator fuer Total Strenght}

\item{use.categorie}{use.categorie=TRUE  M, O, A, 
oder I mit in die Labels ausgeben}

\item{use.points}{use.points = FALSE Punkte statt Text}

\item{type}{Fragetype entwerer vollstaendig (5) oder gekuerzt (3)}

\item{umcodieren}{logical False}

\item{rm_Q}{Remove Q Kategorien Q enfernen Anzahl an erlaubten Qs}

\item{rm_I}{Remove I Kategorien I enfernen Anzahl an erlaubten Is}

\item{methode}{eie sind die Items geornet}

\item{vars_func}{Welche Items sind die Funktionalen}

\item{vars_dysfunc}{Welche Items sind die Dys-Funktionalen}

\item{X, grouping}{intern uebergabe der daten parweisen beides sind data.frames}

\item{x}{Kano-Objekt}
}
\value{
Liste mit:

value: data.frame mit der Kano-Kodierung
scors: Scors,
data= data,
molten=molten,
formula= fm,
removed=Errorrs,
N=nrow(data),
Attributes= c("Must-be","One-dimensional", "Attractive","Indifferent","Reverse", "Questionable"),
answers
}
\description{
Transformiert Kano-Fragebogen zu Kano-Kodierung
http://www.eric-klopp.de/texte/angewandte-psychologie/18-die-kano-methode
}
\details{
M O A I R Q Heufigkeit

max Category

M>O>A>I  max Category mit Hirarchie M Wichtiger als O usw.
also wen der Unterschied zwischen den zwei am hoechsten gelisteten Atributen
zwei Kategorien gleich ist,  5% Schwelle, dann gilt die Regel M>O>A>I

Total Strength als zweite Masszahl gibt an wie hoch der Anteil an bedeutenden
Produktmerkmalen ist.

Category Strength ist eine Masszahl die die angibt ob eine Anforderung nur in
eine Kategorie gehoert

CS plus	Indx Positiv  CS.plus=  (A+O)/(A+O+M+I)

CS minus	Index Negativ CS.minus= (O+M)/(A+O+M+I)

Chi-Test	Eigendlich unsinn Testet ob Verteilung von M, A, O und I gleich ist

Fong-Test Vergleich der zwei Haeufigsten-Kategorien gegenueber der Gesammtzahl
Ergebnis ist entweder ein signifikente oder ein nicht signifikente Verteilung.

\subsection{Basis-Faktoren (Mussfaktoren)}{
 Basis-Merkmale (M...Mustbe) werden vom Kunden Vorausgesetzt schaffen
 unzufriedenheit wenn sie nicht vorhanden sind.
 }
  \subsection{Leistungs- Faktoren}{
 Leistungs-Merkmale (O...One-dimensional) werden vom Kunden verlangt
 }
 \subsection{Begeisterung-Faktoren}{
 Begeisterungs-Merkmale (A...Attractive) Kunde rechnet nicht damit hebt das
 Produkt vom Konkurenten ab.
 }
 \subsection{Unerhebliche- Faktoren }{
 Unerhebliche-Merkmale (I...Indifferent) werden vom Kunden Ignoriert.
 }
}
