% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kano.R
\name{Kano}
\alias{Kano}
\alias{Kano.data.frame}
\alias{Kano.formula}
\alias{Kano_Auswertung}
\alias{kano_plot2}
\title{Analyze Kano type items.}
\usage{
Kano(x, ...)

\method{Kano}{data.frame}(x, na.action = na.pass, ...)

\method{Kano}{formula}(x, data, subset, na.action = na.pass, ...)

Kano_Auswertung(x, caption = "", note = "", formula = x$formula,
  data = x$molten[-1], digits = options()$stp4$apa.style$prozent$digits[1],
  include.n = TRUE, include.percent = TRUE, include.total = TRUE,
  include.test = TRUE, rnd_output = TRUE, ...)

kano_plot2(x, main = "", xlim = c(0, 1), ylim = c(0, 1), mar = c(0, 1,
  2, 1), legend.position = list(x = "right", y = NULL), legend.title = NULL,
  my.lines = "circle", col.bg = "gray95", col = NA, txt.bg = list(m =
  "M", i = "I", o = "O", a = "A"), cex.bg = 12, jitter = TRUE, cex = 1,
  cex.items = cex * 1, cex.lab = cex * 1.07, cex.legend = cex * 1.1,
  ylab = "Zufriedenheitsstiftung (CS+)",
  xlab = "Unzufriedenheitsstiftung (CS-)", center.axis = FALSE,
  use.labels = TRUE, use.total.strength = TRUE, use.categorie = TRUE,
  use.points = FALSE, ...)
}
\arguments{
\item{x, data}{Daten mit/oder formula}

\item{...}{nicht benutzte Argumente}

\item{na.action}{NA's entfernen odere behalten default ist na.pass()}

\item{subset}{an Formula_Data}

\item{caption, note}{Ueberschrift}

\item{formula, data}{intern daten aus Kano-Objekt}

\item{digits}{Nachkommastellen}

\item{include.n}{Anzahl}

\item{include.percent}{Prozent}

\item{include.total}{N und Total}

\item{include.test}{Fong und Chie-Test}

\item{rnd_output}{Intern fuer Plot bei FALSE ausgabe als Zahl}

\item{main}{Header}

\item{mar, xlim, ylim}{Plot Limits}

\item{legend.position, legend.title}{legend.position = list(x="right", y=NULL) 
Beim Streudiagram die Legende bei Barplot  scale=list}

\item{my.lines}{Grafik Element Halbkreis fuer Indiferent}

\item{col}{Farbe der Punkte}

\item{txt.bg, cex.bg, col.bg}{Hintergrund mit den Grossen Bchstaben}

\item{jitter}{Rauschen bei Uberlappung}

\item{cex, cex.items, cex.lab, cex.legend}{Schriftgroesse}

\item{ylab, xlab}{x-y Beschrifung}

\item{center.axis}{center.axis=FALSE}

\item{use.labels}{use.labels=FALSE  Zahlen oder Text als 
Beschriftung bei den Streudiagrammen}

\item{use.total.strength}{use.total.strength=TRUE  groese der Schrift 
als Indikator fuer Total Strenght}

\item{use.categorie}{use.categorie=TRUE  M, O, A, 
oder I mit in die Labels ausgeben}

\item{use.points}{use.points = FALSE Punkte statt Text}

\item{type}{Fragetype entwerer vollstaendig (5) oder gekuerzt (3)}

\item{umcodieren}{logical False}

\item{rm_Q}{Remove Q Kategorien Q enfernen Anzahl an erlaubten Qs}

\item{rm_I}{Remove I Kategorien I enfernen Anzahl an erlaubten Is}

\item{methode}{weie sind die Items geornet}

\item{vars_func}{Welche Items sind die Funktionalen}

\item{vars_dysfunc}{Welche Items sind die Dys-Funktionalen}

\item{X, grouping}{intern uebergabe der daten parweisen beides sind data.frames}

\item{...}{}

\item{x}{Kano-Objekt}
}
\value{
Liste mit:

data: data mit der Kano-Kodierung

scors:  Scors sind eine Alternative Codierung zum zweck der transformierung zu einer metrischen Skala 

molten: Daten-Lang

formula= fm,
removed=Errorrs,
N=nrow(data),
Attributes= c("Must-be","One-dimensional", "Attractive","Indifferent","Reverse", "Questionable"),
answers
}
\description{
Transformiert Kano-Fragebogen zu Kano-Kodierung
}
\details{
http://www.eric-klopp.de/texte/angewandte-psychologie/18-die-kano-methode
https://de.wikipedia.org/wiki/Kano-Modell

\subsection{M Basis-Faktoren (Mussfaktoren)}{
 Basis-Merkmale (\strong{M}ustbe) werden vom Kunden Vorausgesetzt schaffen
 unzufriedenheit wenn sie nicht vorhanden sind.
 }
  \subsection{O Leistungs- Faktoren}{
 Leistungs-Merkmale (\strong{O}ne-dimensional) werden vom Kunden verlangt
 }
 \subsection{A Begeisterung-Faktoren}{
 Begeisterungs-Merkmale (\strong{A}ttractive) Kunde rechnet nicht damit hebt das
 Produkt vom Konkurenten ab.
 }
 \subsection{I Unerhebliche- Faktoren }{
 Unerhebliche-Merkmale (\strong{I}ndifferent) werden vom Kunden ignoriert.
 }
 \subsection{R Rueckweisende- Faktoren }{
 Ablehnende-Merkmale (\strong{R}) werden vom Kunden abgelehnt. Fuehren bei Vorhandensein zu Unzufriedenheit, bei Fehlen jedoch nicht zu Zufriedenheit.
 }

\strong{Kodierung}

Das würde mich sehr freuen (1)
Das setze ich voraus (2)
Das ist mir egal (3)
Das könnte ich in Kauf nehmen (4)
Das würde mich sehr stören (5)



func=1 & dysf=1 = "Q"

func=1 & dysf=2 = "A"

func=1 & dysf=3 = "A"

func=1 & dysf=4 = "A"

func=1 & dysf=5 = "O"

func=2 & dysf=1 = "R"

func=2 & dysf=2 = "I"

func=2 & dysf=3 = "I"

func=2 & dysf=4 = "I"

func=2 & dysf=5 = "M"

func=3 & dysf=1 = "R"

func=3 & dysf=2 = "I"

func=3 & dysf=3 = "I"

func=3 & dysf=4 = "I"

func=3 & dysf=5 = "M"

func=4 & dysf=1 = "R"

func=4 & dysf=2 = "I"

func=4 & dysf=3 = "I"

func=4 & dysf=4 = "I"

func=4 & dysf=5 = "M"

func=5 & dysf=1 = "R"

func=5 & dysf=2 = "R"

func=5 & dysf=3 = "R"

func=5 & dysf=4 = "R"

func=5 & dysf=5 = "Q"

M O A I R Q Heufigkeit

max Category

M>O>A>I  max Category mit Hirarchie M Wichtiger als O usw.
also wen der Unterschied zwischen den zwei am hoechsten gelisteten Atributen
zwei Kategorien gleich ist,  5% Schwelle, dann gilt die Regel M>O>A>I

Total Strength als zweite Masszahl gibt an wie hoch der Anteil an bedeutenden
Produktmerkmalen ist.

Category Strength ist eine Masszahl die die angibt ob eine Anforderung nur in
eine Kategorie gehoert

CS plus	Indx Positiv  CS.plus=  (A+O)/(A+O+M+I)

CS minus	Index Negativ CS.minus= (O+M)/(A+O+M+I)

Chi-Test	Eigendlich unsinn Testet ob Verteilung von M, A, O und I gleich ist

Fong-Test Vergleich der zwei Haeufigsten-Kategorien gegenueber der Gesammtzahl
Ergebnis ist entweder ein signifikente oder ein nicht signifikente Verteilung.
}
