% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kano.R
\name{Kano}
\alias{Kano}
\alias{KANO2}
\alias{Kano}
\alias{Kano_Auswertung}
\title{Analyze Kano type items.}
\usage{
KANO2(X, grouping = NULL, type = 5, umcodieren = FALSE, rm_Q = 10000,
  rm_I = 10000, methode = 1, vars_func = NULL, vars_dysfunc = NULL,
  data = grouping, subset = NULL, na.action = NULL, ...)

Kano(X, grouping = NULL, type = 5, umcodieren = FALSE, rm_Q = 10000,
  rm_I = 10000, methode = 1, vars_func = NULL, vars_dysfunc = NULL,
  data = grouping, subset = NULL, na.action = NULL, ...)

Kano_Auswertung(data, formula,
  digits = options()$stp4$apa.style$prozent$digits[1], prop.table = TRUE,
  ..., var_names = c("N", "Total", "M", "O", "A", "I", "R", "Q",
  "max Category", "M>O>A>I", "Total Strength", "Category Strength", "CS plus",
  "CS minus", "Chi-Test", "Fong-Test"))
}
\arguments{
\item{X}{data.frame mit parweisen Kano Fragen}

\item{grouping}{Dataframe mit Gruppen}

\item{type}{Fragetype entwerer vollstaendig (5) oder gekuerzt (3)}

\item{umcodieren}{logical False}

\item{rm_Q}{Remove Q Kategorien Q enfernen Anzahl an erlaubten Qs}

\item{rm_I}{Remove I Kategorien I enfernen Anzahl an erlaubten Is}

\item{methode}{eie sind die Items geornet}

\item{vars_func}{Welche Items sind die Funktionalen}

\item{vars_dysfunc}{Welche Items sind die Dys-Funktionalen}

\item{...}{\itemize{
   \item \code{type=c(1,2)}  Streudiagram (1) oder Balken (2)
   \item \code{groups= NULL} wenn ein Dataframe uebergeben dan sind das die Gruppen
   \item \code{legend.position = list(x="right", y=NULL) } - Beim Streudiagram die
   Legende bei Barplot \code{scale=list()}
   \item \code{txt.bg=list(m="M",i= "I", o="O", a="A")} Beim Streudiagram der
   Hintergrundtext
   \item \code{jitter=TRUE} Beim Streudiagram Verrauschen der Items (wegen besserer
   lesbarkeit)
   \item \code{col.bg="gray95"}
   \item \code{col = NA}
   \item \code{cex=0.70}
   \item \code{cex.items=cex*1}

   \item \code{cex.lab=cex*1.07}
   \item \code{cex.legend=cex*1.1}
   \item \code{ylab="Zufriedenheitsstiftung (CS+)"}
   \item \code{xlab="Unzufriedenheitsstiftung (CS-)"}
   \item \code{CS.minus = "CS minus"}
   \item \code{center.axis=FALSE}
   \item \code{use.labels=FALSE  Zahlen oder Text als Beschriftung bei den Streudiagrammen}
   \item \code{use.total.strength =TRUE  groese der Schrift als Indikator fuer Total Strenght}
   \item \code{use.categorie=TRUE  M, O, A, oder i mit in die Labels ausgeben }

}}
}
\value{
data.frame mit der Kano-Kodierung
}
\description{
Transformiert Kano-Fragebogen zu Kano-Kodierung
http://www.eric-klopp.de/texte/angewandte-psychologie/18-die-kano-methode

M O A I R Q Hoeufigkeit

max Category

M>O>A>I  max Category mit Hirarchie M Wichtiger als O usw.
also wen der Unterschied zwischen den zwei am hoechsten gelisteten Atributen
zwei Kategorien gleich ist,  5% Schwelle, dann gilt die Regel M>O>A>I

Total Strength als zweite Masszahl gibt an wie hoch der Anteil an bedeutenden
Produktmerkmalen ist.

Category Strength ist eine Masszahl die die angibt ob eine Anforderung nur in
eine Kategorie gehoert

CS plus	Indx Positiv  CS.plus=  (A+O)/(A+O+M+I)

CS minus	Index Negativ CS.minus= (O+M)/(A+O+M+I)

Chi-Test	Eigendlich unsinn Testet ob Verteilung von M, a, O und I gleich ist

Fong-Test Vergleich der zwei Haeufigsten-Kategorien gegenueber der Gesammtzahl
Ergebnis ist entweder ein signifikente oder ein nicht signifikente Verteilung.


 \subsection{Basis-Faktoren (Mussfaktoren)}{
 Basis-Merkmale (M...Mustbe) werden vom Kunden Vorausgesetzt schaffen
 unzufriedenheit wenn sie nicht vorhanden sind.
 }
  \subsection{Leistungs- Faktoren}{
 Leistungs-Merkmale (O...One-dimensional) werden vom Kunden verlangt
 }
 \subsection{Begeisterung-Faktoren}{
 Begeisterungs-Merkmale (A...Attractive) Kunde rechnet nicht damit hebt das
 Produkt vom Konkurenten ab.
 }
 \subsection{Unerhebliche- Faktoren }{
 Unerhebliche-Merkmale (I...Indifferent) werden vom Kunden Ignoriert.
 }
}
\examples{
library(stp5)
 Start("")
 head(DF<-GetData("   Geschlecht f1 f2 f3 f4 f5 f6 f7 f8 f9 f10 f11 f12 f13 f14 f15 f16 f17 f18
 1           w  1  1  1  2  1  3  1  4  1   5   5   1   4   1   5   2   5   1
 2           w  2  1  2  2  2  3  2  4  2   5   2   5   1   3   2   5   2   5
 3           m  3  1  3  2  3  3  3  4  3   5   5   1   4   1   5   2   5   1
 4           m  4  1  4  2  4  3  4  4  4   5   5   1   4   1   5   2   5   1
 5           w  5  1  5  2  5  3  5  4  5   5   5   1   4   1   5   2   5   1
 6           w  NA NA NA NA NA NA NA NA NA  NA  NA  NA  NA  NA  NA  NA  NA  NA
 7           m  2  5  1  5  2  5  1  5  1   5   2   5   1   5   1   5   2   5
 8           w  2  4  2  5  1  3  1  3  2   5   3   3   1   3   1   4   1   3
 9           m  2  4  2  5  2  3  1  3  2   5   1   3   1   3   2   4   3   3
 10          m  2  5  1  5  1  4  1  5  2   5   1   4   1   5   2   5   1   3
 11          w  1  5  2  5  1  4  1  4  1   5   1   4   2   5   2   5   1   4
 12          m  2  5  2  5  2  5  2  4  1   5   1   3   1   4   2   5   3   3
 13          w  2  5  2  5  3  3  1  5  2   5   1   5   1   5   3   3   3   3
 14          m  2  5  1  5  1  5  2  5  2  NA   1   5   1   5   2   5   1   5
 15          w  1  4  2  5  1  3  2  5  1   5   1   3   1   4   2   5   1   3
 16          w  1  4  2  5  2  5  2  5  1   5   1   4   2   5   2   5   1   3
 17          w  1  5  2  5  1  5  1  5  2   5   1   4   1   5   2   5   1   4
 18          w  1  5  2  5  2  5  2  5  1   5   1   2   1   5   1   5   3   2
 19          m  1  5  2  5  2  5  2  5  2   5   1   4   2   5   2   5   1   3
 20          w  2  5  2  5  2  5  2  5  2   5   1   3   2   5   2   5   1   3
 21          w  2  5  2  5  1  5  2  5  2   5   1   5   2   5   2   5   1   3
 22          m  1  3  2  5  1  3  2  5  2   5   1   3   2   5   1   3   1   3
 23          w  1  5  2  5  3  3  2  5  2   5   1   4   2   5   2   5   1   4
 24          w  2  4  2  5  2  3  2  4  2   5   3   3   2   4   2   4   1   3
 25          m  2  4  1  5  1  4  2  4  1   5   1   3   3   3   2   4   1   3
 26          w  1  5  1  5  1  3  1  5  1   5   1   3   1   5   1   5   3   3
 27          w  1  5  2  5  3  3  1  4  2   4   1   3   1   3   3   3   5   1
 28          w  2  5  2  5  1  4  2  5  1   5   1   3   2   4   2   5   4   1
 29          w  2  5  2  5  2  4  2  4  2   5   1   4   2   4   1   5   1   4
 30          m  1  5  2  5  1  3  1  3  1   4   1   3   1   3   1   3   1   3
 31          m  1  3  2  5  1  4  1  4  3   3   5   2   2   4  NA  NA   1   3
 "))

 DF<- upData(DF,  labels=c(f1="f1.motiviert"
                           ,f3="f2.anpasst"
                           ,f5="f3.programm"
                           ,f7="f4.info"
                           ,f9="f5.text"
                           ,f11="f6.reagieren"
                           ,f13="f7.text"
                           ,f15="f8.Arzt"
                           ,f17="f9.Telefonanruf"
 ))

#  match(DF$f1, 1) & match(DF$f2, 5)


 kano_res <- KANO2(DF[, -1], grouping = DF[1],
       # type = 5,
       # umcodieren = FALSE,
       # rm_Q = 999,
       # rm_I = 999
 )


 kano_res <- KANO2( .~ Geschlecht, DF)


 names(kano_res)
 #"value"      "scors"      "data"       "molten"     "formula"  "removed"    "Attributes" "answers"

 APA2(kano_res, caption = "Ueber Objekt")
 APA2(value ~ variable, kano_res, caption = "Ueber formula")
 APA2(value ~ variable + Geschlecht , kano_res, caption = "Ueber formula +Gechlecht")

#Kano_Auswertung( kano_res$molten, kano_res$formula)


library(stp25plot)
 Kano_plot(value ~ variable + Geschlecht, kano_res,
     main = "kano-analysis",
     ylab = "Prozent",
     type = 2, col = 2:7,
     prop.table = TRUE)

 Kano_plot(value ~ variable + Geschlecht, kano_res,
     main = "kano-analysis",
     center.axis = TRUE)

#   windows(8, 8)

 End()

}
