% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{Ordnen}
\alias{Ordnen}
\alias{Ordnen.default}
\alias{Ordnen.anova}
\alias{Ordnen.aov}
\alias{Ordnen.lm}
\alias{Ordnen.glm}
\alias{Ordnen.merModLmerTest}
\alias{Ordnen.lmerModLmerTest}
\alias{Ordnen.polr}
\title{Ordnen (Tidy)}
\usage{
Ordnen(x, ...)

\method{Ordnen}{default}(x, ...)

\method{Ordnen}{anova}(x, ...)

\method{Ordnen}{aov}(x, include.eta = TRUE, include.sumsq = TRUE,
  include.meansq = FALSE, ...)

\method{Ordnen}{lm}(x, include.b = TRUE, include.se = TRUE,
  include.beta = FALSE, include.ci = FALSE, include.r = TRUE,
  include.ftest = FALSE, ci.level = 0.95, ...)

\method{Ordnen}{glm}(x, digits = 2, include.b = TRUE, include.se = TRUE,
  include.b.ci = FALSE, include.ci = TRUE, include.odds = TRUE,
  include.odds.ci = include.ci, include.test = "lrt", include.r = TRUE,
  include.pseudo = include.r, include.loglik = TRUE, ci.level = 0.95, ...)

\method{Ordnen}{merModLmerTest}(x, include.b = TRUE, include.se = TRUE,
  include.beta = FALSE, include.ci = FALSE, include.odds = FALSE,
  ci.level = 0.95, ...)

\method{Ordnen}{lmerModLmerTest}(x, include.b = TRUE, include.se = TRUE,
  include.beta = FALSE, include.ci = FALSE, include.odds = FALSE,
  ci.level = 0.95, ...)

\method{Ordnen}{polr}(x, digits = 2, include.b = TRUE, include.se = TRUE,
  include.ci = FALSE, include.odds = TRUE, ...)
}
\arguments{
\item{x}{Objekt}

\item{...}{weitere Objekte nicht benutzt}

\item{include.eta}{Eta Quadrat}

\item{include.sumsq, include.meansq}{Quadrat- Summen}

\item{include.b}{Estimate}

\item{include.se}{SE Standardfehler}

\item{include.beta}{standartisiertes beta}

\item{include.ci, ci.level}{95%-CI mit ci-Level}
}
\value{
ein data.frame-Ojekt oder eine Liste von data.frames. Im Attribut N sind die Stichprobengroesse
und notes
}
\description{
Diese Funktion meine Version von tidy

ANOVA

Regression (linear)

lmerTest::lmer returns an object of class 'lmerModLmerTest' (previously 'merModLmerTest') 
to clarify that 'lmerModLmerTest' extends  'lmerMod' – not 'merMod'. The merMod class includes generalized 
and nonlinear mixed models and lmerTest is only designed for linear mixed models.
}
