% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndr.R
\name{rndr_}
\alias{rndr_}
\alias{rndr_median}
\alias{rndr_median_quant}
\alias{rndr_median_range}
\alias{rndr_mean}
\alias{rndr_mean_range}
\alias{rndr_ods}
\alias{rndr_percent}
\alias{rndr_P}
\alias{rndr_Stars}
\alias{rndr_corr}
\alias{rndr_r}
\alias{rndr_r2}
\alias{rndr_r2pseudo}
\alias{rndr_df}
\alias{rndr_CI}
\alias{rndr_ods_CI}
\alias{rndr_mean_CI}
\alias{rndr_F}
\alias{rndr_T}
\alias{rndr_H}
\alias{rndr_BP}
\alias{rndr_DW}
\alias{rndr_W}
\alias{rndr_U}
\alias{rndr_shapiro}
\alias{rndr_lm}
\alias{rndr_X}
\alias{rndr_Chisq}
\alias{rndr_Chisq_stars}
\alias{rndr_fischer}
\alias{rndr_gfi_cfa}
\alias{rndr_agfi_cfa}
\alias{rndr_rmsea_cfa}
\alias{rndr_Chisq_cfa}
\alias{rndr_rmsea_cfa}
\alias{rndr_cfi_cfa}
\alias{rndr_nfi_cfa}
\alias{symbol_chi2}
\title{Render}
\usage{
rndr_(...)

rndr_median(m, iqr, digits = NULL, ...)

rndr_median_quant(x, digits = NULL, ...)

rndr_median_range(m, iqr, mn, mx, digits = NULL, ...)

rndr_mean(m, s, digits = NULL, ...)

rndr_mean_range(m, s, mn, mx, digits = NULL, ...)

rndr_ods(x, digits = 2)

rndr_percent(x, n = NULL, ...)

rndr_P(p, include.symbol = TRUE, include.bracket = FALSE)

rndr_Stars(p)

rndr_corr(x, p, df)

rndr_r(x, include.symbol = TRUE)

rndr_r2(x, ...)

rndr_r2pseudo(x, ...)

rndr_df(df1, df2 = NULL)

rndr_CI(ci, digits = 2, sep = options()$stp25$apa.style$sep_element,
  sep_1 = options()$stp25$apa.style$brackets[1],
  sep_2 = options()$stp25$apa.style$brackets[2])

rndr_ods_CI(ci, digits = 2, sep = options()$stp25$apa.style$sep_element,
  sep_1 = options()$stp25$apa.style$brackets[1],
  sep_2 = options()$stp25$apa.style$brackets[2])

rndr_mean_CI(m, ci, digits)

rndr_F(F_val, df1, df2, p = NULL)

rndr_T(F_val, df1, p = NULL)

rndr_H(F_val, df1, p = NULL)

rndr_BP(F_val, df1, p = NULL)

rndr_DW(F_val, df1, p = NULL)

rndr_W(F_val, p = NULL)

rndr_U(F_val, p = NULL)

rndr_shapiro(F_val, p = NULL)

rndr_lm(F_val, df1, df2, p, r2, ar2)

rndr_X(x, df1, df2 = NULL, p = NULL)

rndr_Chisq(x, df, p)

rndr_Chisq_stars(x, p)

rndr_fischer(x, p)

rndr_gfi_cfa(x)

rndr_agfi_cfa(x)

rndr_rmsea_cfa(x)

rndr_Chisq_cfa(x, df = 1)

rndr_rmsea_cfa(x)

rndr_cfi_cfa(x)

rndr_nfi_cfa(x)

symbol_chi2(output = stp25output:::which_output())
}
\arguments{
\item{...}{alles an Format2()}

\item{m}{Mittelwert}

\item{digits}{Nachkommastellen}

\item{x}{Obkekt oder vektor}

\item{s, iqr}{SD,IRQ  (ein Wert)}

\item{n}{Anzahl}

\item{p}{p-Wert}

\item{include.symbol, include.bracket}{Formatierungs Optionen}

\item{df, df1, df2}{Freiheitsgrade}

\item{ci}{Vektor mit zwei Werten}

\item{sep, sep_1, sep_2}{intern seperator}

\item{F_val}{Objekt aus einem Test}

\item{output}{nur intern HTML oder Konsole}

\item{percent, style}{Formatierung als Prozent oder als Prozent(Anzahl)}

\item{percentage_str, null_percent_sign}{Formatierungs Optionen}
}
\value{
Character-String
}
\description{
Formatiere von Zahlen nach dem APA-Style ( American Psychological Association ).

rndr_percent()
Percentages are also most clearly displayed in parentheses with no decimal places:
Nearly half (49%) of the sample was married.

Prozent

rndr_corr
Correlations are reported with the degrees of freedom (which is N-2)
in parentheses and the significance level:
The two variables were strongly correlated, r(55) = .49, p < .01.

F-Wert \code{rndr_F()}

T-Wert \code{rndr_T()}
T Tests are reported like chi-squares, but only the degrees of freedom are
 in parentheses. Following that, report the t statistic (rounded to two decimal places)
  and the significance level.
 There was a significant effect for gender, t(54) = 5.43, p < .001, with men receiving higher scores than women.

CFA Confirmatorische Faktoranalyse

Backhaus Multivariate Analysemethoden 11 AuflageSeite 383
GIF Goodness-of-Fit-Index >=.9

AGIF Adjusted-Goodness-of-Fit-Index

SRMR

Chisq_cfa:  Moosbrugger, Kelava 2012 Testtheorie 2. Auflage Seite 339
CHISQ Chi-Quadrat/df 0,2, 3

RMSEA Root-Mean-Square-Error of Approximation 0, 0.050, 0.080

CFI Comparative-Fit-Index .970-1.00, .950-.969

NFI Normed-Fit-Index .950-1.00, .900-.949
}
\details{
see:
http://winvector.github.io/APAsig/APAsig.html
 http://my.ilstu.edu/~jhkahn/apastats.html
     https://web2.uconn.edu/writingcenter/pdf/Reporting_Statistics.pdf
}
\examples{
rndr_(1.234, 3)

 rndr_mean_range(1,2,3,4)
  rndr_mean (1,2 )
  rndr_median_range(1,2,3,4)
  rndr_median(1,2)

#-- rndr_percent ------------

rndr_percent(c(.2568, 99, 0.02568), c(4, 569, 25), digits = 1)
rndr_percent(10, 3, F, 2)
x <- c(.2568, 99, 0.02568)
n = c(4, 569, 25)
percent = TRUE
digits = options()$stp25$apa.style$prozent$digits[1]
percentage_str = options()$stp25$apa.style$prozent$percentage_str
style = options()$stp25$apa.style$prozent$style
null_percent_sign = options()$stp25$apa.style$prozent$null_percent_sign

rndr_percent(x, percentage_str = "Prozent")

x <- xtabs(~ med + g, hyper)
n <- table(x)
x <- prop.table(n)
rndr_percent(x, n)


hkarz$LAI <- factor(hkarz$lai, 0:1, c("pos", "neg"))
hkarz$Tzell <- cut(hkarz$tzell, 3, c("low", "med", "hig"))
x <- xtabs(~ LAI + Tzell, hkarz)

n <- as.matrix(x)
x <- as.matrix(prop.table(x) * 100)

rndr_percent(x, n)

rndr_percent(data.frame(x), data.frame(n))


rndr_P(c(1,.25,.08,0.05,0.01,0.0001))
rndr_corr(-.548369,0.005896,55)
  rndr_CI(matrix(c(NA, 1:10, NA), ncol=2))
rndr_T(25.45, 45, .0045)
rndr_H(25.45, 45, .0045)
#capture.output(Hmisc::spearman2(pauli~g, data = rechentest))
 
rndr_gfi_cfa(c(1,.9,.89))
rndr_agfi_cfa(c(1,.9,.89))
rndr_rmsea_cfa(c(1,.9,.89))
rndr_Chisq_cfa(c(0,2,3,2.01,3.4))
rndr_rmsea_cfa(c(0, 0.050, 0.080, .051, .081) )
rndr_cfi_cfa(c(.970,1.00, .950-.969,.8))
rndr_nfi_cfa(c(.950, 1.00, .900,  .949))
}
\keyword{internal}
