% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabelle.R, R/intern_apa2_tests.R
\name{Tabelle}
\alias{Tabelle}
\alias{Tabelle2}
\alias{Tabelle.default}
\alias{Tabelle.lm}
\alias{Describe}
\alias{errate_statistik3}
\alias{corr_tabel_X}
\alias{conTest}
\alias{catTest}
\title{Tabelle}
\usage{
Tabelle(...)

Tabelle2(...)

\method{Tabelle}{default}(..., formula = NULL, fun = NULL, type = c("2",
  "1", "auto", "freq", "mean", "median", "ci", "multiresponse", "cohen.d",
  "effsize", "freq.ci", "describe"), caption = "Charakteristik", note = "",
  digits = NULL, APA = FALSE, test = FALSE, na.action = na.pass,
  exclude = NA, include.n = TRUE, include.nr = FALSE,
  include.total = FALSE, include.test = test, include.p = TRUE,
  include.sig.star = FALSE, corr_test = "pearson",
  cor_diagonale_up = TRUE, max_factor_length = 35, order = FALSE,
  decreasing = FALSE, useconTest = FALSE, normality.test = FALSE)

\method{Tabelle}{lm}(x, ..., caption, note)

Describe(x, ...)

errate_statistik3(..., type = NULL, caption = "", note = "",
  digits = NULL, test = FALSE, na.action = na.pass, exclude = NA,
  include.n = TRUE, include.nr = FALSE, include.total = FALSE,
  include.test = test, include.p = TRUE, include.sig.star = FALSE,
  corr_test = "pearson", cor_diagonale_up = TRUE, max_factor_length = 35,
  order = FALSE, decreasing = FALSE, useconTest = FALSE,
  normality.test = FALSE)

corr_tabel_X(..., type = c("pearson", "spearman"), exclude = NA, subset,
  na.action = na.pass, cor_diagonale_up = TRUE, sig.star = TRUE,
  p.value = FALSE, include.mean = FALSE, include.n = TRUE)

conTest(fml, data, test_name = TRUE)

catTest(fml, data, include.test = "chisq.test")
}
\arguments{
\item{...}{Die auszuwertenden Variablen  sex, age="mean", usw}

\item{formula}{An dcast Gruppe ~ .id ist zum Zeilen und Spalten vertauschen}

\item{fun}{Eigene Function am Berechne}

\item{type}{1 oder 2 1 ist kurzes Format 2 int lang}

\item{caption, note}{Wir in attribute gespeichert Ueberschrift und Note.}

\item{digits}{Kommastellen}

\item{APA}{APA2 versuion}

\item{test, include.test, corr_test, useconTest, normality.test}{Signifikanz Test}

\item{na.action, exclude}{An Formula nict benutzt}

\item{include.n, include.nr}{Ausgabe von Anzahl}

\item{include.total}{Ausgabe vonTotal}

\item{include.p, include.sig.star}{p-Werte, Sternchen}

\item{cor_diagonale_up}{Korrelation}

\item{max_factor_length}{Fehler Abfangen bei langen Factoren}

\item{order, decreasing}{noch nicht Implementiert}

\item{.data}{Daten data.frame}

\item{type}{welche Auswertung}
}
\value{
Tabelle: data.frame oder list mit data.frame
Tabelle2: HTML
}
\description{
Einfache deskriptive Tabelle die in medizinischen Arbeiten
verwendet werden.
Die Funktion arbeitet Intern mit \code{aggregate} bzw. mi  berechne.default().

Tabelle2:  html-Output Tabelle(...) %>% Output()

Hilfsfunktion fuer Tabellen

Hilfsfunktion fuer Tabellen

conTest Hilfsfunktion fuer Tabellen

conTest Hilfsfunktion fuer Tabellen
}
\examples{

library(dplyr)
library(tidyr)
#names(varana)
varana2 <- varana \%>\%
  gather(Zeit, Merkfgk, m1:m4) \%>\%
  mutate(Zeit = factor(Zeit, Cs(m1, m2, m3 , m4), Cs(t0, t1, t2, t3))) \%>\%
  Label(Merkfgk = "Merkfaehigkeit")

Tabelle(Merkfgk ~ Zeit, varana2)

varana2 \%>\% Tabelle(Merkfgk, by =  ~ Zeit)
varana \%>\% Tabelle(m1, m2, m3 , m4)

#- Achtung hier wird eine Liste Ausgegeben
varana \%>\% Tabelle(
  4:7,
  by =  ~ geschl,
  fun = function(x)
    c(
      n = length(na.omit(x)),
      m = mean(x),
      sd = sd(x)
    )
)

varana2 \%>\% Tabelle2(Merkfgk, by=~ Zeit)
prepare_data2(m1+m2~m3 , varana)
corr_tabel_X(~m1+m2, varana)
corr_tabel_X(m1+m2~m3, varana)
corr_tabel_X(m1+m2~m3|geschl, varana)
}
\author{
Wolfgang Peter
}
